+++ src/lib.rs
@@ -59,7 +59,7 @@ macro_rules! aw {
 }
 
 pub trait Day {
-    fn part1(&self) -> i32;
-    fn part2(&self) -> i32;
+    fn part1(&self) -> usize;
+    fn part2(&self) -> usize;
     fn fmt_result(&self) -> String;
 }
